Added another optimization for dict::key lookups
The key object caches the index at which the key/value pair was found
the last time, and checks that index first the next time. This will
frequently succeed, since it's common for multiple dictionaries to have
the same set of keys (look at 1000people.json for example.)

Future optimization: If the comparison fails, we can use the sign (-1
or +1) to limit the ensuing search to either before or after that
index.