Removed FleeceDocument.mm from library target
It's not needed by client code that doesn't use Obj-C (e.g. LiteCore)
but won't be dead-stripped by the linker. Also, it prevents 32-bit
compilation due to using ivars in the @implementation, which is an
issue for working with Xamarin.

Obviously anyone who needs it can just compile the source code into
their own targets.