Reimplemented alloc_slice to be more efficient
No longer based on std::shared_ptr. Instead the heap buffer is
allocated with an extra 4 bytes for a ref-count at the start. The
‘buf’ field points just after the ref-count.

The drawback is that you can no longer cheaply adopt a malloc block as
an alloc_slice, or detach an alloc_slice and use it as a malloc block.