Fixed a Path bug with PersistentSharedKeys
The static Path evaluator didn’t work correctly if evaluating a Dict
that has numeric keys that don’t yet appear in the SharedKeys instance.
It needs to refresh the SharedKeys from storage to pick up the new keys.
Fortunately Dict already has a getter that does that.

Was causing couchbase/couchbase-lite-ios#1587