Added C++ conveniences to C API
This isn’t crazy. If Fleece is built as a dynamic library, only the C
API should be exported, since C++ name mangling is compiler dependent
and C++ interfaces aren’t very binary-stable. But a client using that
dylib may be written in C++, so it’ll be using the Fleece C API. This
commit makes the FL___ structs more like C++ objects.