Avoid casting out-of-range float values to integers
The encoder checks if a float/double value can be written as an integer,
to save space. It did this by testing "n == (int64_t)n". But the behavior
of this is considered undefined if n is out of the range of int64_t,
and Clang's undefined-behavior sanitizer flags it.
I've changed the test so it doesn't do this.