Converted mutable classes to use public/wrapper C++ API
i.e. the API in FleeceCpp.hh, not the one in Fleece.hh
This allows the mutable stuff to be implemented/used in situations
where only the Fleece C API is available.

In the process I fixed some issues with encoding Obj-C to Fleece,
with the result that in Objective-C to make a custom class Fleece-
encodable you now implement -fl_encodeToFLEncoder: not
-fl_substitutedEncoding. This is easier to use and more efficient.