Mutability changes; MContext is now public and subclassable
MArray and MDict now keep track of (and enforce) their own mutability.

mutableContainers is no longer an attribute of the Context,
rather of each container.

MContext can now be subclassed, for clients that need to store extra
state across the object tree, or that use something other than an
alloc_slice to keep the Fleece data in memory.