Fixed an edge case in Encoder where it’d throw an exception
checkPointerWidths() looks for pointers that will exceed the range
of a narrow Value (65535 bytes), and switches the array/dict to wide
if one is found.
However, it was computing the pointer delta from the current write
position ... but the array/dict header itself will be written before
the pointers are written, which adds 2+ bytes to each delta. This
left an opportunity for a pointer to overflow 65535 bytes when
written.
Fixed this and added a unit test that reproduced the exception and
no longer fails.

Fixes couchbase/couchbase-lite-core#493