Dict delta encoding!
A Dict can now have a parent Dict from which it inherits keys/values.
Only the values that are different need to be written.
MutableDict will, of course, encode itself in this form if it will
save space.
The parent is stored in the first slot of the Dict, as the value, with
a magic key of -2048.