DB is now more durable
DB makes an attempt to guard against file corruption due to a crash
or power loss.
- File is padded to a multiple of 4k bytes.
- The file ends with a trailer structure containing a magic number.
- If a file is opened and its size or trailer aren’t valid, DB will
  jump back to the end of the previous 4k page and look for a trailer
  there, repeating until it runs out of file. If no trailer is found,
  it throws an exception.

This isn’t fully durable yet, but it’s a start. The main gotchas are
that in a power failure the final page might get written but not one
of the pages before, or the final page might get partially written
so the trailer’s magic number is valid but not other data. I think
the fix for this will require a second filesystem flush.