Fixed CMake
Building Fleece itself with `CMake` failed with a cryptic error from
`make`, although it could still be built as a dependency of LiteCore.
The problem was with the executable `fleece` (the command line tool);
it seems that having a target with the same name as the project
caused confusion, or maybe that only happens on case-folding file
systems like HFS…
While I was at it I also added an exported-symbol list to the dynamic
library.