API: Breaking change to mutable setters
Instead of 26 separate Set/Append functions there are now only three:
FLMutableArray_Set, FLMutableArray_Append, FLMutableDictionary_Set.
These return an FLSlot, which can be used with nine new setters to
set the value. For example:
    FLSlot_SetInt(FLMutableArray_Append(a), 17);
In the C++ API, the set() and append() methods return a Slot object
which has an operator= overload for each data type. For example:
    a.append() = 17