Restored compatibility with data predating ‘extern’ flag
There are existing persisted documents with narrow pointers whose
offsets are >32KB. In those pointers, the bit now used for ‘offset’
is set, causing them to be misinterpreted, causing the doc to fail to
validate.
Fixed this by double-checking such failures: If an extern pointer
doesn’t validate, and it’s narrow, and if its interpretation in the
old format makes it a valid pointer into the same doc, then it’s
interpreted as a ‘legacy’ pointer.

Fixes #45