KeyTree: String prefix sharing
If all children of a node share a common prefix, then the node gives
the length of that prefix and the prefix is removed from all children.
This reduces the data size for larger dictionaries, and probably
speeds up lookups a bit due to fewer redundant byte comparisons.

To do: The prefix-length is most often zero, especially in small data
sets, but still takes up one byte. Find a way to eliminate that,
probably by squeezing in a flag bit elsewhere that indicates whether
there is a prefix or not.