Fixed bugs with MutableDict and PersistentSharedKeys
Setting a key in a MutableDict should not add the key to the Shared-
Keys, because if the latter are persistent and this is not during a
transaction, an exception will be thrown. Instead it should be added
in string form.

If that key later gets added to the SharedKeys, the HeapDict needs to
behave properly even though it's storing the key in string form; this
took a few more tweaks to achieve.

Added tests to exercise this.

Found in couchbaselabs/couchbase-lite-C#18