Fix a mistake in using setlocale
I misunderstood that the old locale would be returned, but actually it's the same as what is passed in, and furthermore it's a live pointer so every call to set locale changes it.  Not worth even looking at.  Man pages say that every program starts up in "C" locale, and we only change it here so it's going to be in "C", and then go back to "C"