Regular string literals can now be used efficiently as slices
The slice constructor that takes a char* can now be evaluated at
compile time if the parameter is a string literal. In that case
there's no call to strlen.

Because of that I also took out the "explicit" keyword, so you can
now just use a string constant for a slice parameter instead of
having to use the "_sl" suffix. Hurray!

(Non-literal C strings still work; but of course there's still a
call to strlen() to find the string's length.)