Correct parsing of very large JSON integers
jsonsl doesn't check for overflow when it parses integers, so we
would get bogus values for anything outside the range of a uint64_t or
(for negative numbers) an int64_t. For example, 18446744073709551616
(UINT64_MAX + 1) would parse as 0!

I added some careful parsing functions to NumConversions.hh, which
fail on overflow. JSONConverter now calls these if the numeric string
is long enough that it could overflow, and if it does, it parses it as
a double instead. This will of course lose some precision, but it's
better than converting to something totally wrong, as we were doing.