Fixed Retained::operator= to avoid re-entrant destruction
Retained::operator= used to release the old value before storing the
new value. If this destructed the old value, its destructor would still
see itself as the value of the Retained instance. In some cases this
could lead to a double-dispose crash, generally if a reference cycle
was being broken.

Fixed this by changing the order of operations so that the new value
is stored before releasing the old one.

At the same time I improved the `moveRef` function, putting the whole
implementation of assignment into it. This shrinks the call site
significantly, important since assignments happen a lot.

See https://issues.couchbase.com/browse/CBL-1281