Remove gmtime_r and localtime_r usage
They don't exist on Windows and the alternatives cannot handle dates before the Epoch.  Instead use the date.h library which provides a fast efficient calendar implementation on top of the chrono library.  Mainly gmtime_r and localtime_r were used as tools to figure out the offset from UTC for a given timestamp, but the date.h library does it in a more readable way without losing much, if any, efficiency (most of the chrono library is constexpr and can be reduced to integral math at compile time)