Added FLEncoder_Snip
This allows you to encode two documents at once, where the second one
reuses data from the first. The second document has to be loaded with
the first one as its "extern".
This will be used in LiteCore to write a document's current revision
properties, plus the other available revisions, without having to
load the latter from the database unless necessary.