Retained: Disabled dangerous implicit conversion to T*
There is no longer an implicit conversion from an rvalue Retained<T>
to T*, because this conversion is easy to do by mistake in contexts
where it results in a dangling pointer to a freed T.

Please see the comment starting at RefCounted.hh:153 for details.

Also added assignments from Retained<T> to RetainedConst<T>,
which is of course valid.