Fixed Android 11 incompatibility with ValueSlot ptr tagging
The prior size optimization in ValueSlot used the high byte of the
`_pointer` field to distinguish between an actual pointer and an
inline 7-bit Value; it assumed if the upper byte is zero then it's a
pointer.

This isn't true anymore in Android: "Starting in Android 11, for 64-
bit processes, all heap allocations have an implementation defined tag
set in the top byte of the pointer on devices with kernel support for
ARM Top-byte Ignore (TBI)."

Switched to using the _low_ byte as a tag, with a value of 0xFF
denoting inline, and anything else being a pointer. This works because
any Value* we store there must have at least one zero bit in the
LSB; see the comment in ValueSlot.hh for details.