CBL-2273: Export global Fleece symbols
This is a confusing situation but if compiled in a static library these directives don't have any effect directly.  The issue happens when the static library is then included in a shared library which then wants to export the symbols in its interface.  Both the declaration and definition need to be exported on Windows, but you cannot retroactively compile an exported definition.  This compiled export will carry through and apply to the eventual shared library.

NOTE: Warning, an export list will override this and make this symbol invisible again so be sure to also include it in export lists if you want to use it.