Fixed a bad implicit conversion with mutable operator[]
A boolean test of a mutable array/dict property, like
`if (dict["key"])`, would always evaluate to true; there was an
incorrect implicit conversion from `keyref` to `FLSlot` that made
the test evaluate to a non-NULL FLSlot pointer.

Removing the FLSlot conversion operator fixed it I couldn't deduce
any need for the operator, and removing it didn't break anything in
LiteCore or CBL-C.

I did add a setNull() method, for parity with Slot.

Fixes #105