Split Fleece.h and Fleece.hh into multiple headers
Fleece.h is now an umbrella header that #includes the new headers:
  FLBase.h, FLCollections.h, FLDeepIterator.h, FLDoc.h, FLEncoder.h,
  FLJSON.h, FLKeyPath.h, FLMutable.h, FLValue.h
There is also FLExpert.h, which is stuff that clients are unlikely to
use, or too dangerous, or semi-deprecated.

On the C++ side I’ve only split out the expert stuff, into Expert.hh.

Also I renamed Base.h to CompilerSupport.h, to keep it from getting
mixed up with FLBase.h (not to mention Base.hh in LiteCore.)