CBL-2867: Fixed incorrect iteration of HeapDict due to key mixup
If a HeapDict's overlay map stores the string form of a key, but its
parent Dict uses the int form, iteration will find _both_ forms of the
key. This is obviously bad news.

Fixed this by making sure that when a new entry is added to the
overlay map, it uses the same form of key as found in the parent Dict.
(It used to just look the key up in the SharedKeys, but that's not
enough in the edge case where the in-memory SharedKeys are not up to
date and the parent Dict uses a key that's not in memory.)