Fixed a bug in FLEncoder_LastValueWritten, FLEncoder_Snip, FLEncoder_WriteValueAgain
The internal enum type Encoder::PreWrittenValue had a badly-chosen
"none" value of 0, which could easily occur as a valid value.

This meant Encoder::lastValueWritten() would return a false negative
if the last value written was the first value (at offset 0.)

This in turn meant Encoder::snip() would return nullslice in this
case.

This in turn broke LiteCore's VectorRecord in this case.

Further improvements to writeValueAgain

- Added public constant kFLNoWrittenValue
- writeValueAgain returns false if the last value can't be rewritten