Add a separate property, isCacheable, to SharedKeys.
If cacheable, Dict::get(key &keyToFind) will cache the string key to integer key and enhance the performance of the ensuing gets. This is good if the doucment is not changed.

For PersistentSharedKey, the library manages the property: it is cacheable only if it is not in a transaction when Dict::get is called.

For SharedKeys, the client manages it. If it anticipates the sharedKeys may outlive the underlying document changes, it should call SharedKeys::disableCaching().