slice: buf & size should not be initialized in their declaration
This reverts a problematic change made in October in commit 170d7170.

pure_slice's buf and size fields are, for historical reasons, public
but read-only. The read-only nature is enforced by declaring them
'const'.

Recently while refactoring I moved the default values of those fields
into the declaration, i.e. adding `= 0` and `= nullptr`. This is
correct code, but it turned out to confuse some static analysis
tools, which now assume that these fields will _always_ be 0 or NULL.
As a result, in CLion I'm seeing IDE warning markers that some code
is unreachable or some expressions are always true or always false.
These are just noise, really, but they're annoying.

It's easy enough to fix this regression by taking out the
initializers on the variable declarations and moving them back to the
default constructor.