JSONConverter can now parse plain scalars
The jsonsl parser we use follows an older version of the JSON spec
which didn't allow a document to be a scalar value, only an array or
object.

I've run into this issue a few times lately -- for example, JSON
Schema considers `true` or `false` a valid schema, but we can't
parse it, so a few tests in their test suite break.

I made a small change to JSONConverter to detect when the input isn't
an array or object, and wrap it in `[...]` so that jsonsl can parse
it. (But meanwhile it ignores the outer array when parsing.)