Fix: SharedKeys::encode() doesn't work on empty key
A weird edge case that showed up when running the JSON Schema test
suite, which includes some JSONPointer paths with empty-string
components.

The bug is that `_table.find(str)` finds the entry for the empty
string, but the following test `entry.key != nullslice` fails
because an empty slice compares as == nullslice since they're both
empty. The proper way to test whether a slice is not nullslice is
to use `if(entry.key)`, which doesn't work here because
`__usuallyTrue` requires a boolean, or `entry.key.buf != nullptr`.